справлены замечания: проект модульный, структура пересобрана, тесты проходят
