+++ tests/conftest.py
@@ -1,4 +1,5 @@
-import os
 import sys
+import os
 
+# Добавляет src в sys.path
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'src')))
